backend/requirements.txt
asgiref==3.8.1
Django==5.1.2
sqlparse==0.5.1
google-generativeai==0.8.3
djangorestframework==3.15.2